fix: test with drops
