+++ ledger/ledger_test.go
@@ -19,7 +19,6 @@ package ledger
 import (
 	"bytes"
 	"context"
-	"database/sql"
 	"errors"
 	"fmt"
 	"math/rand"
@@ -2280,7 +2279,7 @@ func TestLedgerReloadTxTailHistoryAccess(t *testing.T) {
 	// reset tables and re-init again, similary to the catchpount apply code
 	// since the ledger has only genesis accounts, this recreates them
 	err = l.trackerDBs.Batch(func(ctx context.Context, tx store.BatchScope) error {
-		arw, err := tx.CreateAccountsWriter()
+		arw, err := tx.MakeAccountsWriter()
 		if err != nil {
 			return err
 		}
@@ -2339,21 +2338,7 @@ func TestLedgerReloadTxTailHistoryAccess(t *testing.T) {
 
 	// drop new tables
 	// reloadLedger should migrate db properly
-	err = l.trackerDBs.Batch(func(ctx context.Context, tx store.BatchScope) error {
-		var resetExprs = []string{
-			`DROP TABLE IF EXISTS onlineaccounts`,
-			`DROP TABLE IF EXISTS txtail`,
-			`DROP TABLE IF EXISTS onlineroundparamstail`,
-			`DROP TABLE IF EXISTS catchpointfirststageinfo`,
-		}
-		for _, stmt := range resetExprs {
-			_, err0 := tx.ExecContext(ctx, stmt)
-			if err0 != nil {
-				return err0
-			}
-		}
-		return nil
-	})
+	err = l.trackerDBs.ResetToV6Test(context.Background())
 	require.NoError(t, err)
 
 	err = l.reloadLedger()
@@ -2639,18 +2624,7 @@ func TestLedgerMigrateV6ShrinkDeltas(t *testing.T) {
 	cfg.MaxAcctLookback = shorterLookback
 	store.AccountDBVersion = 7
 	// delete tables since we want to check they can be made from other data
-	err = trackerDB.Batch(func(ctx context.Context, tx *sql.Tx) error {
-		if _, err := tx.ExecContext(ctx, "DROP TABLE IF EXISTS onlineaccounts"); err != nil {
-			return err
-		}
-		if _, err := tx.ExecContext(ctx, "DROP TABLE IF EXISTS txtail"); err != nil {
-			return err
-		}
-		if _, err = tx.ExecContext(ctx, "DROP TABLE IF EXISTS onlineroundparamstail"); err != nil {
-			return err
-		}
-		return nil
-	})
+	err = trackerDB.ResetToV6Test(context.Background())
 	require.NoError(t, err)
 
 	l2, err := OpenLedger(log, dbName, inMem, genesisInitState, cfg)
